backend/requirements.txt
altair>=5.0.0
annotated-types>=0.6.0
anyio>=4.0.0
attrs
blinker
cachetools
certifi
charset-normalizer
click>=8.0.0
ephem>=4.1
fastapi>=0.109.0
filelock
fsspec
gitdb
GitPython
h11
hf-xet
huggingface-hub
idna
Jinja2
joblib
jsonschema
jsonschema-specifications
MarkupSafe
mpmath
narwhals
neo4j==5.16.0
networkx>=3.0
numpy>=1.24.0
packaging
pandas>=2.0.0
pillow>=10.0.0
protobuf
pyarrow
pydantic>=2.0.0
pydantic_core
pydeck
pyswisseph
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests>=2.31.0
rpds-py
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
six
smmap
starlette
streamlit>=1.30.0
sympy
tenacity
threadpoolctl
tokenizers
toml
--extra-index-url https://download.pytorch.org/whl/cpu
torch

tornado
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn>=0.27.0
matplotlib
reportlab
